rename it to disco.py
