+++ torchtitan/optimizers/__init__.py
@@ -5,7 +5,7 @@
 # LICENSE file in the root directory of this source tree.
 
 from .abstract_disco import AbstractDiSCO  # noqa: F401
-from .distributed_scion import DiSCO  # noqa: F401
+from .disco import DiSCO  # noqa: F401
 from .utils import (  # noqa: F401  # noqa: F401  # noqa: F401
     create_disco_optimizer_kwargs_from_optimizer_config,
     create_disco_param_groups,
